requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
sentence-transformers
streamlit
requests
numpy